Add an expression for small integers

+++ src/main/java/com/jetbrains/jetpad/vclang/core/expr/visitor/ToAbstractVisitor.java
@@ -475,6 +475,6 @@ public class ToAbstractVisitor extends BaseExpressionVisitor<Void, Concrete.Expr
 
   @Override
   public Concrete.Expression visitInteger(IntegerExpression expr, Void params) {
-    return new Concrete.NumericLiteral(null, expr.getInteger());
+    return new Concrete.NumericLiteral(null, expr.getBigInteger());
   }
 }
